\name{internalFunctions}
\Rdversion{1.1}
\alias{AC.node.trans}
\alias{getDI}
\alias{nested.mean.overlap}
\alias{nbConnectingNodes}
\alias{noi}
\alias{sister}
\alias{terminal.sisters}
\alias{thinPNO}

\title{
PHYLOCLIM Internal Functions
}

\description{
This is an internal function in PHYLOCHLIM and not intended to be called by the user.
}