\name{phyloclim-internal}
\alias{AC.node.trans}
\alias{descendants}
\alias{di.enm}
\alias{di.pno}
\alias{import.asc}
\alias{nbConnectingNodes}
\alias{nested.mean.overlap}
\alias{noi}
\alias{sister}
\alias{terminal.sisters}
\alias{thinPNO}

\title{Internal PhyloClim Functions}
\description{Internal \pkg{phyloclim} functions.}
\note{
These are internal functions in \code{PhyloClim} and are not intended to be called by the user.
}
\keyword{internal}
