% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pno.R
\name{pno}
\alias{pno}
\title{Predicted Niche Occupancy Profiles}
\usage{
pno(path_bioclim, path_model, subset = NULL, bin_width = 1,
  bin_number = NULL)
}
\arguments{
\item{path_bioclim}{A character string, giving the path to an ASCII raster
map with environmental data (bioclimatic layer) and ArcGIS-compatible
header (see Details).}

\item{path_model}{A character string, giving the path to the
\bold{directory}, where MAXENT probability surfaces are stored as ASCII
raster maps with ArcGIS-compatible header (see Details).}

\item{subset}{A vector of mode 'character' giving taxon names if only a
subset of the species ENMs in \code{path_model} is to be used}

\item{bin_width}{A numeric, giving the bin interval.}

\item{bin_number}{An integer, giving the number of bins used.}
}
\value{
A data frame of cumulative probalities of suitability for each value
  on an environmental gradient (rows) for the a set of studied taxa (rows).
  The first column gives the mean of each category.
}
\description{
Following the approach of Evans et al. (2009), \code{pno}
  integrates species probability (suitability) distributions (e.g. derived
  with MAXENT) with respect to single climatic (or other environmental)
  variable in order to get predicted niche occupancy (PNO) profiles for each
  species and environmental variable.
}
\details{
The path argument \code{path_model} points to the directory where
  the MAXENT probability distributions must be stored as ASCII maps with an
  \bold{ArcGIS-compatible} header (as obtained with MAXENT output filetype:
  \bold{asc}). 
  
  An example for an ArcGIS-compatible header: \tabular{rll}{ \tab ncols   \tab 958 \cr
  \tab nrows   \tab 376 \cr \tab xllcorner     \tab -73.440304 \cr \tab
  yllcorner     \tab -55.27622 \cr \tab cellsize      \tab 0.0166 \cr \tab
  NODATA_value  \tab -9999 \cr }
}
\references{
Evans, M. E. K., S. A. Smith, R. S. Flynn, and M. J. Donoghue.
  2009. Climate, niche evolution,and diversification of the 'bird-cage
  evening primroses' (\emph{Oenothera}, sections \emph{Anogra} and
  \emph{Kleinia}). \emph{Am. Nat.} \bold{173}: 225-240.
}
\seealso{
\code{\link{pno.weighted.mean}}, \code{\link{niche.overlap}},
  \code{\link{anc.clim}}
}
